requirements.txt
numpy
tqdm>=4.23.0
scipy==1.2
torchsummary>=1.3
matplotlib>=2.2.2
opencv_contrib_python>=4.1.0.25 # >=4.1.0.25
torchvision>=0.3.0
scikit_image>=0.15.0
torchviz>=0.0.1
Pillow>=6.1.0
torch==1.10.1
easydict
prefetch_generator
PyYAML
tensorboard
future
numba
pypng
scikit-image
protobuf==3.20.0
